Add rule tests.
