--- packages/style/tests/conditions.test.ts
@@ -1,78 +1,0 @@
-import ClientRenderer from '../src/client/ClientRenderer';
-import { getRenderedStyles, purgeStyles } from '../src/testing';
-
-describe('Conditions', () => {
-  let renderer: ClientRenderer;
-
-  beforeEach(() => {
-    renderer = new ClientRenderer();
-  });
-
-  afterEach(() => {
-    purgeStyles();
-  });
-
-  describe('media queries', () => {
-    it('supports @media conditions', () => {
-      const className = renderer.renderRule({
-        background: '#000',
-        padding: '15px',
-        '@media (max-width: 600px)': {
-          padding: '15px',
-        },
-        '@media screen and (min-width: 900px)': {
-          padding: '20px',
-        },
-      });
-
-      expect(className).toBe('a b c d');
-      expect(getRenderedStyles('standard')).toMatchSnapshot();
-      expect(getRenderedStyles('conditions')).toMatchSnapshot();
-    });
-
-    it('can be nested in @supports', () => {
-      const className = renderer.renderRule({
-        padding: '15px',
-        '@supports (display: flex)': {
-          '@media (max-width: 600px)': {
-            padding: '15px',
-          },
-        },
-      });
-
-      expect(className).toBe('a b');
-      expect(getRenderedStyles('standard')).toMatchSnapshot();
-      expect(getRenderedStyles('conditions')).toMatchSnapshot();
-    });
-  });
-
-  describe('support queries', () => {
-    it('supports @supports conditions', () => {
-      const className = renderer.renderRule({
-        display: 'block',
-        '@supports (display: flex)': {
-          display: 'flex',
-        },
-      });
-
-      expect(className).toBe('a b');
-      expect(getRenderedStyles('standard')).toMatchSnapshot();
-      expect(getRenderedStyles('conditions')).toMatchSnapshot();
-    });
-
-    it('can be nested in @media', () => {
-      const className = renderer.renderRule({
-        display: 'block',
-        '@media screen and (min-width: 900px)': {
-          '@supports (display: flex)': {
-            display: 'flex',
-          },
-        },
-      });
-
-      expect(className).toBe('a b');
-      expect(getRenderedStyles('standard')).toMatchSnapshot();
-      expect(getRenderedStyles('conditions')).toMatchSnapshot();
-    });
-  });
-});
